+++ Archivos/Cliente/ClienteEntidad.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GrupoA.Prototipo.Archivos;
+namespace GrupoA.Prototipo.Archivos.Cliente;
 internal class ClienteEntidad
 {
     public string CuitCliente { get; set; }
