+++ Archivos/Mercaderias/MercaderiaArchivo.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GrupoA.Prototipo.Archivos;
+namespace GrupoA.Prototipo.Archivos.Mercaderias;
 internal class MercaderiaArchivo
 {
     public static ReadOnlyCollection<MercaderiaEntidad> Mercaderias => mercaderias.AsReadOnly();
