+++ Archivos/Mercaderias/MercaderiasDetalle.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GrupoA.Prototipo.Archivos
+namespace GrupoA.Prototipo.Archivos.Mercaderias
 {
     public class MercaderiasDetalle
     {
