+++ Archivos/Ordenes/OrdenEntrega/OrdenEntregaEntidad.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GrupoA.Prototipo.Archivos;
+namespace GrupoA.Prototipo.Archivos.Ordenes.OrdenEntrega;
 internal class OrdenEntregaEntidad
 {
     public int NroOrdenEntrega { get; set; }
