+++ Archivos/Ordenes/OrdenSeleccion/OrdenSeleccionArchivo.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using GrupoA.Prototipo.Archivos.Estados;
+using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -6,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace GrupoA.Prototipo.Archivos;
+namespace GrupoA.Prototipo.Archivos.Ordenes.OrdenSeleccion;
 internal static class OrdenSeleccionArchivo
 {
     public static ReadOnlyCollection<OrdenSeleccionEntidad> OrdenesSeleccion => ordenesseleccion.AsReadOnly();
