+++ DespachoConTransportista/DespachoConTransportistaModel.cs
@@ -1,4 +1,6 @@
 ﻿using GrupoA.Prototipo.Archivos;
+using GrupoA.Prototipo.Archivos.Estados;
+using GrupoA.Prototipo.Archivos.Stock;
 using GrupoA.Prototipo.RetiroStock;
 using System;
 using System.Collections.Generic;
